feat: update deployment scripts
